refactor: move ArduinosState and ArduinoDomain to core modules
This commit moves the `ArduinosState` data class from `feature/conectionInternet` to `ui/core/models` and `ArduinoDomain` from `domain/conectionInternet` to `domain/core/models`.

This change centralizes these core data models, making them accessible to other features and promoting better code organization.

Affected files:
- `AddArduinoViewModel.kt`: Updated import for `ArduinosState`.
- `AddArduinoScreen.kt`: Updated import for `ArduinosState`.
- `AddArduinoUseCase.kt`: Updated imports for `ArduinoDomain` and `StateObjectDomain`.
