+++ feature/conectionInternet/src/commonMain/kotlin/com/virtualworld/multiplatformiot/feature/conectionInternet/screen/AddArduinoScreen.kt
@@ -5,8 +5,8 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.virtualworld.multiplatformiot.feature.conectionInternet.models.ArduinosState
 import com.virtualworld.multiplatformiot.ui.core.component.ButtonBack
+import com.virtualworld.multiplatformiot.ui.core.models.ArduinosState
 
 @Composable
 fun AddArduinoScreen(
